% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.R
\name{fitModel}
\alias{fitModel}
\title{Title}
\usage{
fitModel(
  .data,
  x = NULL,
  y,
  .newdata = NULL,
  level = "mean",
  regression = "qrf",
  removeZeroInstances = FALSE,
  uniqueResponses = FALSE,
  ntree = 500,
  n_bootstrap = 100,
  k = 5
)
}
\arguments{
\item{.data}{data.table}

\item{x}{character}

\item{y}{character}

\item{.newdata}{vector}

\item{level}{double.}

\item{regression}{character}

\item{removeZeroInstances}{logical}

\item{uniqueResponses}{logical}

\item{ntree}{integer}

\item{n_bootstrap}{integer Number of bootstrap samples for KNN}

\item{k}{integer Number of nearest neighbors for KNN}
}
\value{
vector
}
\description{
Title
}
