% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.R
\name{fitModel}
\alias{fitModel}
\title{Title}
\usage{
fitModel(
  .data,
  variables = NULL,
  response,
  .newdata = NULL,
  level = "mean",
  regression = "qrf",
  removeZeroInstances = FALSE,
  uniqueResponses = FALSE,
  cvmax = 10,
  nbmax = 100
)
}
\arguments{
\item{.data}{data.table}

\item{variables}{character}

\item{response}{character}

\item{.newdata}{vector}

\item{level}{double.}

\item{regression}{character}

\item{removeZeroInstances}{logical}

\item{uniqueResponses}{logical}

\item{cvmax}{integer}

\item{nbmax}{integer}
}
\value{
vector
}
\description{
Title
}
